pyproject.toml
[tool.poetry]
name = "repl_nix_jovianwebsiteflask"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
SQLAlchemy = "^2.0.7"
mysql-connector-python = "^8.0.32"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
